Proper impersonation in scenario migration is displayed in activities panel.
