+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/ScenarioActivityApiHttpService.scala
@@ -393,7 +393,6 @@ class ScenarioActivityApiHttpService(
             date,
             scenarioVersionId,
             sourceEnvironment,
-            sourceUser,
             sourceScenarioVersionId,
             targetEnvironment,
           ) =>
