+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/description/MigrationApiEndpoints.scala
@@ -22,7 +22,8 @@ import pl.touk.nussknacker.security.AuthCredentials
 import pl.touk.nussknacker.ui.api.description.MigrationApiEndpoints.Dtos.{
   MigrateScenarioRequestDto,
   MigrateScenarioRequestDtoV1,
-  MigrateScenarioRequestDtoV2
+  MigrateScenarioRequestDtoV2,
+  MigrateScenarioRequestDtoV3
 }
 import pl.touk.nussknacker.ui.migrations.MigrationService.MigrationError
 import pl.touk.nussknacker.ui.migrations.MigrationService.MigrationError.{
@@ -207,6 +208,20 @@ object MigrationApiEndpoints {
         isFragment: Boolean,
     ) extends MigrateScenarioRequestDto
 
+    @derive(encoder, decoder)
+    final case class MigrateScenarioRequestDtoV3(
+        override val version: Int,
+        sourceEnvironmentId: String,
+        sourceScenarioVersionId: Option[VersionId],
+        processingMode: ProcessingMode,
+        engineSetupName: EngineSetupName,
+        processCategory: String,
+        scenarioLabels: List[String],
+        scenarioGraph: ScenarioGraph,
+        processName: ProcessName,
+        isFragment: Boolean,
+    ) extends MigrateScenarioRequestDto
+
     /*
     NOTE TO DEVELOPER:
 
@@ -274,7 +289,8 @@ object MigrationApiEndpoints {
 
         implicit val migrateScenarioRequestV1Schema: Schema[MigrateScenarioRequestDtoV1] = Schema.derived
         implicit val migrateScenarioRequestV2Schema: Schema[MigrateScenarioRequestDtoV2] = Schema.derived
-//        implicit val migrateScenarioRequestV3Schema: Schema[MigrateScenarioRequestDtoV3] = Schema.derived
+        implicit val migrateScenarioRequestV3Schema: Schema[MigrateScenarioRequestDtoV3] = Schema.derived
+//        implicit val migrateScenarioRequestV4Schema: Schema[MigrateScenarioRequestDtoV4] = Schema.derived
 
         val derived = Schema.derived[MigrateScenarioRequestDto]
         derived.schemaType match {
@@ -286,7 +302,8 @@ object MigrationApiEndpoints {
                 Map(
                   "1" -> SchemaType.SRef(Schema.SName(classOf[MigrateScenarioRequestDtoV1].getSimpleName)),
                   "2" -> SchemaType.SRef(Schema.SName(classOf[MigrateScenarioRequestDtoV2].getSimpleName)),
-//                  "3" -> SchemaType.SRef(Schema.SName(classOf[MigrateScenarioRequestDtoV2].getSimpleName)),
+                  "3" -> SchemaType.SRef(Schema.SName(classOf[MigrateScenarioRequestDtoV3].getSimpleName)),
+//                  "4" -> SchemaType.SRef(Schema.SName(classOf[MigrateScenarioRequestDtoV4].getSimpleName)),
                 )
               )
             )
@@ -298,14 +315,16 @@ object MigrationApiEndpoints {
       implicit val encoder: Encoder[MigrateScenarioRequestDto] = Encoder.instance {
         case v1: MigrateScenarioRequestDtoV1 => v1.asJson
         case v2: MigrateScenarioRequestDtoV2 => v2.asJson
-//        case v3: MigrateScenarioRequestDtoV3 => v3.asJson
+        case v3: MigrateScenarioRequestDtoV3 => v3.asJson
+//        case v4: MigrateScenarioRequestDtoV4 => v4.asJson
       }
 
       implicit val decoder: Decoder[MigrateScenarioRequestDto] = Decoder.instance { cursor =>
         cursor.downField("version").as[Int].flatMap {
           case 1 => cursor.as[MigrateScenarioRequestDtoV1]
           case 2 => cursor.as[MigrateScenarioRequestDtoV2]
-//          case 3     => cursor.as[MigrateScenarioRequestDtoV3]
+          case 3 => cursor.as[MigrateScenarioRequestDtoV3]
+//          case 4     => cursor.as[MigrateScenarioRequestDtoV4]
           case other => throw new IllegalStateException(s"Cannot decode migration request for version [$other]")
         }
       }
