+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/migrate/HttpRemoteEnvironment.scala
@@ -1,0 +1,197 @@
+package pl.touk.nussknacker.ui.process.migrate
+
+import cats.data.EitherT
+import cats.implicits._
+import com.typesafe.scalalogging.LazyLogging
+import io.circe.Decoder
+import io.circe.syntax.EncoderOps
+import org.apache.pekko.actor.ActorSystem
+import org.apache.pekko.http.scaladsl.{Http, HttpExt}
+import org.apache.pekko.http.scaladsl.model._
+import org.apache.pekko.http.scaladsl.model.Uri.{Path, Query}
+import org.apache.pekko.http.scaladsl.model.headers.{Authorization, BasicHttpCredentials, RawHeader}
+import org.apache.pekko.http.scaladsl.unmarshalling.{Unmarshal, Unmarshaller}
+import org.apache.pekko.stream.Materializer
+import pl.touk.nussknacker.engine.api.process.{ProcessName, ScenarioVersion, VersionId}
+import pl.touk.nussknacker.restmodel.scenariodetails.ScenarioWithDetailsForMigrations
+import pl.touk.nussknacker.ui.NuDesignerError
+import pl.touk.nussknacker.ui.NuDesignerError.XError
+import pl.touk.nussknacker.ui.api.description.MigrationApiEndpoints.Codecs.MigrateScenarioRequestDto.encoder
+import pl.touk.nussknacker.ui.api.description.MigrationApiEndpoints.Dtos.{ApiVersion, MigrateScenarioRequestDto}
+import pl.touk.nussknacker.ui.migrations.MigrateScenarioData
+import pl.touk.nussknacker.ui.security.api.{AuthManager, ImpersonatedUserData, ImpersonationSupport, LoggedUser}
+
+import scala.concurrent.{Await, ExecutionContext, Future}
+import scala.concurrent.duration.DurationInt
+
+final case class HttpRemoteEnvironmentConfig(
+    uri: String,
+    user: String,
+    password: String,
+    targetEnvironmentId: String,
+    remoteConfig: StandardRemoteEnvironmentConfig
+)
+
+class HttpRemoteEnvironment(
+    httpConfig: HttpRemoteEnvironmentConfig,
+    val testModelMigrations: TestModelMigrations,
+    val environmentId: String,
+    impersonationSupport: ImpersonationSupport
+)(implicit as: ActorSystem, val materializer: Materializer)
+    extends StandardRemoteEnvironment
+    with LazyLogging
+    with AutoCloseable {
+  override val config: StandardRemoteEnvironmentConfig = httpConfig.remoteConfig
+
+  private val closeTimeout = 10 seconds
+
+  def baseUri: Uri = Uri(httpConfig.uri)
+
+  val http: HttpExt = Http()
+
+  override def close(): Unit = Await.ready(closeAsync(), closeTimeout)
+
+  def closeAsync(): Future[Unit] = http.shutdownAllConnectionPools()
+
+  override def processVersions(processName: ProcessName)(
+      implicit ec: ExecutionContext
+  ): Future[List[ScenarioVersion]] =
+    invokeJson[ScenarioWithDetailsForMigrations](HttpMethods.GET, List("processes", processName.value)).map { result =>
+      result.fold(_ => List(), _.historyUnsafe)
+    }
+
+  override protected def fetchRemoteMigrationScenarioDescriptionVersion(
+      implicit ec: ExecutionContext
+  ): EitherT[Future, NuDesignerError, Int] = {
+    EitherT {
+      invokeJson[ApiVersion](
+        HttpMethods.GET,
+        List("migration", "scenario", "description", "version"),
+        Query.Empty,
+        requestEntity = HttpEntity.Empty
+      ).map(_.map(_.version))
+    }
+  }
+
+  override protected def migrateScenario(migrateScenarioData: MigrateScenarioData)(
+      implicit ec: ExecutionContext,
+      loggedUser: LoggedUser
+  ): Future[Either[NuDesignerError, Unit]] = {
+    val dto = MigrateScenarioData.fromDomain(migrateScenarioData)
+    invokeForSuccess(
+      HttpMethods.POST,
+      List("migrate"),
+      Query.Empty,
+      HttpEntity(dto.asJson.noSpaces),
+      impersonationHeader(loggedUser)
+    )
+  }
+
+  override protected def fetchProcesses(
+      implicit ec: ExecutionContext
+  ): Future[Either[NuDesignerError, List[ScenarioWithDetailsForMigrations]]] = {
+    invokeJson[List[ScenarioWithDetailsForMigrations]](
+      HttpMethods.GET,
+      List("processes"),
+      Query(("isArchived", "false"))
+    )
+  }
+
+  override protected def fetchProcessVersion(name: ProcessName, remoteProcessVersion: Option[VersionId])(
+      implicit ec: ExecutionContext
+  ): Future[Either[NuDesignerError, ScenarioWithDetailsForMigrations]] = {
+    invokeJson[ScenarioWithDetailsForMigrations](
+      HttpMethods.GET,
+      List("processes", name.value) ++ remoteProcessVersion.map(_.value.toString).toList,
+      Query()
+    )
+  }
+
+  override protected def fetchProcessesDetails(names: List[ProcessName])(
+      implicit ec: ExecutionContext
+  ): EitherT[Future, NuDesignerError, List[ScenarioWithDetailsForMigrations]] = EitherT {
+    invokeJson[List[ScenarioWithDetailsForMigrations]](
+      HttpMethods.GET,
+      "processesDetails" :: Nil,
+      Query(
+        ("names", names.map(_.value).mkString(",")),
+        ("isArchived", "false"),
+        ("skipNodeResults", "true"),
+      )
+    )
+  }
+
+  private def request(
+      uri: Uri,
+      method: HttpMethod,
+      request: MessageEntity,
+      headers: Seq[HttpHeader]
+  ): Future[HttpResponse] = {
+    logger.debug("Sending request to remote environment: {} {}", method.value, uri)
+    http.singleRequest(
+      HttpRequest(
+        uri = uri,
+        method = method,
+        entity = request,
+        headers = List(Authorization(BasicHttpCredentials(httpConfig.user, httpConfig.password))) ++ headers
+      )
+    )
+  }
+
+  private def impersonationHeader(loggedUser: LoggedUser): List[HttpHeader] =
+    impersonationSupport.toImpersonatedUserIdentityWithSupportCheck(
+      ImpersonatedUserData(loggedUser.id, loggedUser.username, Set.empty)
+    ) match {
+      case Right(identityString) => List(RawHeader(AuthManager.impersonateHeaderName, identityString))
+      case Left(_)               => Nil
+    }
+
+  private def invoke[T](
+      method: HttpMethod,
+      pathParts: List[String],
+      query: Query = Query.Empty,
+      requestEntity: RequestEntity = HttpEntity.Empty,
+      headers: Seq[HttpHeader]
+  )(f: HttpResponse => Future[T])(implicit ec: ExecutionContext): Future[T] = {
+    val pathEncoded = pathParts.foldLeft[Path](baseUri.path)(_ / _)
+    val uri         = baseUri.withPath(pathEncoded).withQuery(query)
+
+    request(uri, method, requestEntity, headers) flatMap f
+  }
+
+  private def invokeForSuccess(
+      method: HttpMethod,
+      pathParts: List[String],
+      query: Query = Query.Empty,
+      requestEntity: RequestEntity,
+      headers: Seq[HttpHeader]
+  )(implicit ec: ExecutionContext): Future[XError[Unit]] =
+    invoke(method, pathParts, query, requestEntity, headers) { response =>
+      if (response.status.isSuccess()) {
+        response.discardEntityBytes()
+        Future.successful(().asRight)
+      } else {
+        Unmarshaller
+          .stringUnmarshaller(response.entity)
+          .map(error => RemoteEnvironmentCommunicationError(response.status, error).asLeft)
+      }
+    }
+
+  private def invokeJson[T: Decoder](
+      method: HttpMethod,
+      pathParts: List[String],
+      query: Query = Query.Empty,
+      requestEntity: RequestEntity = HttpEntity.Empty
+  )(implicit ec: ExecutionContext): Future[Either[NuDesignerError, T]] = {
+    invoke(method, pathParts, query, requestEntity, headers = Nil) { response =>
+      if (response.status.isSuccess()) {
+        Unmarshal(response.entity).to[T].map(Either.right)
+      } else {
+        Unmarshaller
+          .stringUnmarshaller(response.entity)
+          .map(error => Either.left(RemoteEnvironmentCommunicationError(response.status, error)))
+      }
+    }
+  }
+
+}
