+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/migrate/ProcessModelMigrator.scala
@@ -24,7 +24,6 @@ final case class MigrationResult(process: CanonicalProcess, migrationsApplied: L
           canonicalProcess = process,
           labels = labels,
           increaseVersionWhenJsonNotChanged = true,
-          forwardedUserName = None,
           migrationsApplied = migrations
         )
       }
