+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/server/PekkoRoutesFactory.scala
@@ -10,7 +10,7 @@ import pl.touk.nussknacker.ui.factory.{DomainServices, InfrastructureServices}
 import pl.touk.nussknacker.ui.process.{ConfigScenarioToolbarService, ProcessStateDefinitionService}
 import pl.touk.nussknacker.ui.process.deployment.{DeploymentService => LegacyDeploymentService}
 import pl.touk.nussknacker.ui.process.migrate.{HttpRemoteEnvironment, ProcessModelMigrator, TestModelMigrations}
-import pl.touk.nussknacker.ui.security.api.{AuthenticationResources, LoggedUser}
+import pl.touk.nussknacker.ui.security.api.{AuthenticationResources, ImpersonationSupport, LoggedUser}
 
 import scala.concurrent.ExecutionContext
 
@@ -34,7 +34,8 @@ object PekkoRoutesFactory {
         designerConfig,
         infrastructureServices,
         domainServices,
-        customHttpServiceProviders
+        customHttpServiceProviders,
+        authenticationResources.impersonationSupport
       )
 
       routesWithoutAuthentication =
@@ -47,7 +48,8 @@ object PekkoRoutesFactory {
       designerConfig: DesignerConfig,
       infrastructureServices: InfrastructureServices,
       domainServices: DomainServices,
-      customHttpServiceProviders: Map[String, CustomHttpServiceProvider]
+      customHttpServiceProviders: Map[String, CustomHttpServiceProvider],
+      impersonationSupport: ImpersonationSupport
   ): List[RouteWithUser] = {
     import domainServices._
     import infrastructureServices._
@@ -122,7 +124,8 @@ object PekkoRoutesFactory {
                 .mapValues(new ProcessModelMigrator(_)),
               processingTypeServicesProvider.mapValues(_.scenarioValidator)
             ),
-            designerConfig.environment
+            designerConfig.environment,
+            impersonationSupport
           )
           new RemoteEnvironmentResources(
             remoteEnvironment,
