+++ designer/server/src/test/resources/config/business-cases/basicauth-users.conf
@@ -15,6 +15,11 @@ users: [
     roles: []
   },
   {
+    identity: "remoteUser"
+    password: "remoteUser"
+    roles: ["Writer"]
+  },
+  {
     identity: "allpermuser"
     password: "allpermuser"
     roles: ["AllPermissions"]
