+++ designer/server/src/test/scala/pl/touk/nussknacker/test/processes/WithScenarioActivitySpecAsserts.scala
@@ -60,7 +60,7 @@ trait WithScenarioActivitySpecAsserts
              |  "activities": [
              |    {
              |      "id": "${regexes.looseUuidRegex}",
-             |      "user": "allpermuser",
+             |      "user": "remoteUser",
              |      "date": "${regexes.zuluDateRegex}",
              |      "scenarioVersionId": 1,
              |      "additionalFields": [],
@@ -68,7 +68,7 @@ trait WithScenarioActivitySpecAsserts
              |    },
              |    {
              |      "id": "${regexes.looseUuidRegex}",
-             |      "user": "allpermuser",
+             |      "user": "remoteUser",
              |      "date": "${regexes.zuluDateRegex}",
              |      "scenarioVersionId": 2,
              |      "additionalFields": [
