+++ designer/server/src/test/scala/pl/touk/nussknacker/test/utils/domain/ProcessTestData.scala
@@ -449,7 +449,7 @@ object ProcessTestData {
       edges = List.empty
     )
 
-    UpdateScenarioCommand(scenarioGraph, comment.map(_.content), Some(List.empty), None)
+    UpdateScenarioCommand(scenarioGraph, comment.map(_.content), Some(List.empty))
   }
 
   def validProcessWithFragment(
