+++ designer/server/src/test/scala/pl/touk/nussknacker/test/utils/domain/ScenarioToJsonHelper.scala
@@ -20,7 +20,6 @@ object ScenarioToJsonHelper {
         scenarioGraph,
         comment = None,
         scenarioLabels = Some(List.empty),
-        forwardedUserName = None
       ).asJson
 
   }
