+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/MigrationApiHttpServiceBusinessSpec.scala
@@ -45,7 +45,7 @@ class MigrationApiHttpServiceBusinessSpec
         .statusCode(200)
         .body(
           "version",
-          equalTo[Int](2)
+          equalTo[Int](3)
         )
     }
   }
@@ -56,6 +56,7 @@ class MigrationApiHttpServiceBusinessSpec
         given()
           .when()
           .basicAuthAllPermUser()
+          .impersonateRemoteUser()
           .jsonBody(validRequestData)
           .post(s"$nuDesignerHttpAddress/api/migrate")
           .Then()
@@ -85,6 +86,7 @@ class MigrationApiHttpServiceBusinessSpec
         given()
           .when()
           .basicAuthAllPermUser()
+          .impersonateRemoteUser()
           .jsonBody(prepareRequestJsonDataV1(exampleProcessName.value, exampleGraph, isFragment = false))
           .post(s"$nuDesignerHttpAddress/api/migrate")
           .Then()
@@ -115,6 +117,7 @@ class MigrationApiHttpServiceBusinessSpec
           )
           .when()
           .basicAuthAllPermUser()
+          .impersonateRemoteUser()
           .jsonBody(validRequestDataV2)
           .post(s"$nuDesignerHttpAddress/api/migrate")
           .Then()
@@ -143,6 +146,7 @@ class MigrationApiHttpServiceBusinessSpec
           )
           .when()
           .basicAuthAllPermUser()
+          .impersonateRemoteUser()
           .jsonBody(prepareRequestJsonDataV1(exampleProcessName.value, exampleGraphV2, isFragment = false))
           .post(s"$nuDesignerHttpAddress/api/migrate")
           .Then()
@@ -169,6 +173,7 @@ class MigrationApiHttpServiceBusinessSpec
       given()
         .when()
         .basicAuthAllPermUser()
+        .impersonateRemoteUser()
         .jsonBody(requestDataWithInvalidScenarioName)
         .post(s"$nuDesignerHttpAddress/api/migrate")
         .Then()
@@ -184,6 +189,7 @@ class MigrationApiHttpServiceBusinessSpec
         )
         .when()
         .basicAuthAllPermUser()
+        .impersonateRemoteUser()
         .jsonBody(validRequestData)
         .post(s"$nuDesignerHttpAddress/api/migrate")
         .Then()
@@ -200,6 +206,7 @@ class MigrationApiHttpServiceBusinessSpec
           )
           .when()
           .basicAuthAllPermUser()
+          .impersonateRemoteUser()
           .jsonBody(validRequestDataForFragmentV2)
           .post(s"$nuDesignerHttpAddress/api/migrate")
           .Then()
@@ -228,6 +235,7 @@ class MigrationApiHttpServiceBusinessSpec
           )
           .when()
           .basicAuthAllPermUser()
+          .impersonateRemoteUser()
           .jsonBody(prepareRequestJsonDataV1(validFragment.name.value, exampleFragmentGraphV2, isFragment = true))
           .post(s"$nuDesignerHttpAddress/api/migrate")
           .Then()
@@ -256,6 +264,7 @@ class MigrationApiHttpServiceBusinessSpec
         given()
           .when()
           .basicAuthAllPermUser()
+          .impersonateRemoteUser()
           .jsonBody(validRequestDataForFragment)
           .post(s"$nuDesignerHttpAddress/api/migrate")
           .Then()
@@ -281,6 +290,7 @@ class MigrationApiHttpServiceBusinessSpec
         given()
           .when()
           .basicAuthAllPermUser()
+          .impersonateRemoteUser()
           .jsonBody(prepareRequestJsonDataV1(validFragment.name.value, exampleFragmentGraph, isFragment = true))
           .post(s"$nuDesignerHttpAddress/api/migrate")
           .Then()
