+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/notifications/NotificationServiceTest.scala
@@ -333,7 +333,6 @@ class NotificationServiceTest
         canonicalProcess = sampleScenario,
         processingType = Streaming.stringify,
         isFragment = false,
-        forwardedUserName = None
       )
     writeProcessRepository
       .saveNewProcess(action)(TestFactory.adminUser())
