+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/deployment/DeploymentServiceSpec.scala
@@ -908,7 +908,6 @@ class DeploymentServiceSpec
       canonicalProcess = canonicalProcess,
       processingType = "streaming",
       isFragment = false,
-      forwardedUserName = None
     )
     writeProcessRepository.saveNewProcess(action).map(_.value.processId).map(ProcessIdWithName(_, processName))
   }
@@ -924,7 +923,6 @@ class DeploymentServiceSpec
       canonicalProcess = canonicalProcess,
       processingType = "streaming",
       isFragment = true,
-      forwardedUserName = None
     )
 
     writeProcessRepository.saveNewProcess(action).map(_.value.processId).map(ProcessIdWithName(_, processName))
