+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/newdeployment/DeploymentRepositorySpec.scala
@@ -47,7 +47,6 @@ class DeploymentRepositorySpec
         canonicalProcess = ProcessTestData.validProcess,
         processingType = Streaming.stringify,
         isFragment = false,
-        forwardedUserName = None
       )
     )(TestFactory.adminUser())
     .dbioActionValues
