+++ security/src/main/scala/pl/touk/nussknacker/ui/security/api/ImpersonationSupport.scala
@@ -6,15 +6,25 @@ sealed trait ImpersonationSupport {
       impersonatedUserIdentity: String
   ): Either[ImpersonationNotSupported.type, Option[ImpersonatedUserData]]
 
+  def toImpersonatedUserIdentityWithSupportCheck(
+      userData: ImpersonatedUserData
+  ): Either[ImpersonationNotSupported.type, String]
+
 }
 
 abstract class ImpersonationSupported extends ImpersonationSupport {
   def getImpersonatedUserData(impersonatedUserIdentity: String): Option[ImpersonatedUserData]
+  def toImpersonatedUserIdentity(userData: ImpersonatedUserData): String
 
   override final def getImpersonatedUserDataWithSupportCheck(
       impersonatedUserIdentity: String
   ): Either[ImpersonationNotSupported.type, Option[ImpersonatedUserData]] =
     Right(getImpersonatedUserData(impersonatedUserIdentity))
+
+  override def toImpersonatedUserIdentityWithSupportCheck(
+      userData: ImpersonatedUserData
+  ): Either[ImpersonationNotSupported.type, String] =
+    Right(toImpersonatedUserIdentity(userData))
 
 }
 
@@ -23,6 +33,10 @@ object NoImpersonationSupport extends ImpersonationSupport {
   override final def getImpersonatedUserDataWithSupportCheck(
       impersonatedUserIdentity: String
   ): Either[ImpersonationNotSupported.type, Option[ImpersonatedUserData]] = Left(ImpersonationNotSupported)
+
+  override def toImpersonatedUserIdentityWithSupportCheck(
+      userData: ImpersonatedUserData
+  ): Either[ImpersonationNotSupported.type, String] = Left(ImpersonationNotSupported)
 
 }
